Update actions footer listview asset log reasons
